Spotted some incorrect details about setting processModel
I was looking to try to set my minIoThreads and it appears that the processModel is only settable in the Machine.config by default (Read the documentation linked in https://msdn.microsoft.com/en-us/library/vstudio/7w2sway1(v=vs.100).aspx ). 
The documentation indicated otherwise. In revelation of this, It may be an idea to suggest setting the value via code first!.