Tests: enhance DeslaveGoesToPrimary
This doesn't resolve the issue, but adds several confirmations and more information. At the moment we're not recognizing :6380 as a slaved instance though it clearly is from the outside. There's more to dig into on recognition here. This just improves the test quite a bit, before it was throwing in the middle without much info.