Clarify the documentation for IDatabase(Async).HashSet(Async)
The documentation for the overload of IDatabase(Async).HashSet(Async)
that takes multiple fields/values was somewhat ambiguous. It could have
been read to mean that fields not set as a part of the HashSetAsync are
deleted rather than preserved. The documentation for HMSET was clearer,
so this borrows some of the verbiage from the Redis documentation into
the documentation for HashSetAsync.