Tests: Fix DNS failures
So this one is fun. Windows caches DNS, including a miss. The second test runner will fail synchronously in the Socket.BeginConnect() call due to the DNS looku failing immediately (no need to wait!). This beaks our assumptions of it failing asynchronously in these tests.

If this resolves the issue, we'll put a wrapper in the .BeginConnect() call to capture and recognize the SocketException "No such host is known" and return the RedisConnectionException users expect.