Tests: make profiling handle heartbeat
This is to ensure assumptions are correct. I believe what's happening if we're re-fetching the tie breaker key on a background interval which is landing in the profiler on the multiplexer. This likely the same symptom is affecting several tests. For example, LowAllocationEnumerable.

If this is correct, we should likely find a way to exclude the tie breaker (and any heartbeat activities) from the profiler...or give the profiler an option, but this would require changing the interface (bad) or we add info to a profiled command to indicate. Or we decide this is proper behavior (profiling everything) and we move the tests off DB 0 or something so we can easily exclude the tiebreaker GET.