Add UNLINK as .KeyDelete() automatic usage underneath (#910)
* Add UNLINK as .KeyUnlink()

Submitting as a PR so test as reuable, but maybe we want this as an overload via some enum on .KeyDelete() instead? I'm good either way, but getting the ball rolling on #909 here.

* Add feature check for UNLINK

Ah right, 4.0.0+ only...

* Move UNLINK behind KeyDelete() and always call it if available

We still weren't testing Delete(Async)() in general, so tests added there. I've confirmed this is running DEL against 3.x and UNLINK against 5.x.

* Move GetDeleteCommand after the methods (consistent with rest of the file)