Fix builds for all packages on Linux
Note that we're only building StackExchange.Redis and NRediSearch in the CI build so this wasn't apparent as an issue, but it's also easily remedied.

Cleaning up the xunit MyGet feed as well since it was an intermittent issue earlier in testing this...and turns out we don't use any betas or need them anymore, so why not simplify anyhow.