Fix builds for all packages on Linux (#1316)
Note that we're only building `StackExchange.Redis` and `NRediSearch` in the CI build so this wasn't apparent as an issue, but it's also easily remedied. Cleaning up the xUnit MyGet feed as well since it was an intermittent issue earlier in testing this...and turns out we don't use any betas or need them anymore, so why not simplify anyhow.

Overall, this is a simpler alternative to #1313 and more build simplification/cleanup:

Total changes:
- Moves to 3.x build SDK (just to update)
  - Moves to `Visual Studio 2019` AppVeyor build image, for 3.x support (removes need to install SDK)
  - Rolls forward to latest version
- Moves to `Build.csproj` for building
  - Simplifies `build.ps1`
- Adds `NRediSearch` to the test suite (on Linux in CI, since Windows is against 3.x)
  - Adds `RedisFeatures.Module` (against v 4.0.0) to guard tests against
  - Makes `NRediSearch.Tests` reference `StackExchange.Redis.Tests` and support test skipping
- Makes copyright year range dynamic for the library
- Moves `LangVersion` to latest and in one place (root `Directory.Build.props`)
- Uses `Microsoft.NETFramework.ReferenceAssemblies` to build `net4x` on Linux
- Uses `IsPackable` to default to excluded then reverses to `true` only for the `src/` directory
- Removes `xUnit` MyGet feed, since we're on stable builds now
- Adds new files (and corrects old) in `.sln`
- Removes `LibTargetFrameworks` to simplify since it's less complex given the above changes

cc @mgravell for eyes - this should be all set now!