Exceptions: remove "Unspecified/" leader
I've seen this confusing people in issues when it's just not relevant at all. If there is an address familiy specified, we'll see it. If there's not, there's no point in "Unspecified/" being prepended. This is something .NET Does from DNSEndPoint, not us.

Also adds a test for ExceptionFactory.Timeout to make sure our format is good.