Move NRediSearch off :6379 and Docker-only
This moves NRediSearch to :6385 as its own instance using the official image in Docker instead. Going to play to play with getting this up on Appveyor though.

Instead of loading our module at all, we juse use the image and go boom if the module is missing. No more server mismatches.