Builds & Local: Move NRedisSearch to :6385 and Docker (#1489)
This moves NRedisSearch to docker-compose, and also moves AppVeyor Unbuntu builds to use docker-compose. Unfortunately we can't do this on Windows due to how the build VMs are setup. If we're comfortable doing Docker-only (we didn't want to jump in there), we could cleanup the `.sh` files afterwards.

Additionally, this removes the `redisearch.so` module from our repo and the code to install it (only in tests). Since this broke with every server mismatch, using their official Docker image that'll always work is much saner and simpler.

Fixes NRedisSearch test failures in #1488, but also `master` branch.