Fix a terrible bug. Voila, tests pass more reliably! Summary of changes so far - A 'data-structure-refactor' that changes the `_backlog` queue to support writing concurrently with reading, which can reduce both read and write thread blocking.
I originally wanted to do this as a way to reduce contention between threads adding to the backlog - which was secondary and mainly happens due to pre-existing contention on the write lock (which I haven't tried to eliminate). This was motivated by having too-many-threads logs that I saw while testing some other changes I'm testing for `PhysicalBridge`.

I haven't been able to fully reduce contention between competing reads and competing writes, but by reducing contention between reads and writes, I think therefore this now requires lower max worker threads in general, and makes thread-blocking less likely in total, and worker thread usage lower overall.

The CLR's `ConcurrentQueue` is used because it is mostly non-blocking, and I was hoping it would eliminate needs for `lock()` statement on writes in enqueue scenarios, and even reads.

However, I found that for overall correctness of the backlog management, we are still required to use `lock()` statement on it in deqeueing scenarios, so that we can avoid e.g. dequeueing a message which was not the one we saw with `peek()` (for timeout processing).

And also in order prevent races where `StartProcessingBacklog()` could fail to be called when the queue is not empty, we end up needing to add synchronization between enqueue, and the beginning and end of backlog processing.