Add StateChange handlers
This exposes the physical bridge's connection's state change handlers up on the server endpoint, allowing us to use it for task awaiting or anything else. It's intentionally an Action<T> based mechanism here because ReconfigureAsync runs many places and we have many reentrancy points to the path - ensuring a clean dispose model here in a naive way.

Overall, we're waiting on initial connections (INCLUDNG AUTH) to complete before firing off anything else to an endpoint, like gathering tiebreakers.

This restructures ReconfigureAsync slightly to break the existing race apart (see #1759) and additionally won't send a dual tracer on initial connection as a byproduct (we were blindly sending a second before, the dependency in the race case vs. AUTH).