Idea for #1759 - Resolving race in AUTH connection reconfigurations (#1762)
See notes in #1759 as well.

Overall what's happening here is that the `GetServerEndPoint` on the first connection pass in `ReconfigureAsync` off the multiplexer is asynchronously completed. In most cases, we can append a tracer (`PING` or `ECHO`) to the priority message path and issue them across the wire. They'll come as the next commands and complete in order naturally. However, in the `AUTH` case specifically, we have pending commands that need to come first on that physical connection/socket. By skipping the queue and writing directly to the output stream in the `AUTH` case, what can happen (depending on latency) in an order of:

1. `GetServerEndPoint`
2. Socket connection _begins_ (not completes)
3. Trace (`PING`) is issued
4. `PING` is written to the output pipe before the `AUTH` coming from the handshake is
5. Connection begins handshake phase after physical is established...but at this point we're about to die

When the `PING` gets there, we'll get a `NOAUTH` (properly) back from Redis and kill the connection from bad state.

The race is an obvious problem upon digging, a more subtle one is we're blindly sending a tracer from `ReconfigureAsync` in _all_ cases. However, this is duplicitous in the initial connection case (because we'll send one during the handshake). So instead of sending a tracer on the first path, we can await the handshake completion instead.

The overall change here is awaiting the initial connection or check of existing in the current await path, via some new event exposure around connection state changes, and then splitting tie breakers into another pass after the endpoints are connected. This eliminates the command pipe issuance race during the initial connection phase. We're also normalizing a lot of the logging and removing a sync write path since `AutoConfigure` can be async these days.

Note: `TryConnect`/`GetConnection` were the same method - consolidated for clarity while chasing this down.