Tests: don't fail on *reporting*
In GitHub CI if a person doesn't have permissions (to this repo) then the CI fails from trying to *report* test results. It's fine if tests fail and blow us up as intended, but inability to report them shouldn't. We don't see pretty test results when this happens, but every fork won't be red for no reason either.