Revert tracer flags change
I misunderstood intent of the tracer in this situation - we'll reconfigure if needed on failure, but do not wish to wait on an existing connection check (we assume it's okay). Delaying here is what was failing MultiMaster often due to connection re-use.

Also: only complete the pending handlers for interactive on disconnects as well.