Fixes for #1762 & Don't fail on *reporting* (#1776)
In GitHub CI if a person doesn't have permissions (to this repo) then the CI fails from trying to *report* test results. It's fine if tests fail and blow us up as intended, but inability to report them shouldn't. We don't see pretty test results when this happens, but every fork won't be red for no reason either.

In removing the fire and forget flags from the primary tracer path to await it, I've realized that write path _won't actually complete_ until the buffer fires, creating a deadlock (we should re-evaluate that path, give we write _but don't observe the result_). The tracers purpose here is to ensure the connection is still working, as a sort of keep-alive, mechanisms elsewhere will handle failure and reconnect if so determined (e.g. via `ReconfigureIfNeeded` paths). In short: we want to just fire off a tracer to subsequently check the state of the world. This change restores that behavior.

Overall changes:
- Don't error for non-repo members on test result additions (build infrastructure, not library)
- More connection logging
- Restore `CommandFlags.FireAndForget` to base `GetTracerMessage`
- Only fire task handlers for `OnConnectedAsync` for interactive disconnections (was still doing subscriptions)
- DRY down to `CompletePendingConnectionMonitors`
- Convert `WriteDirectAsync_Awaited` to a local method (cleans up the namespace and normalizes)
- Fix `Cluster` logging (to help a flaky test for `SimpleProfiling`, though it's a hard repro atm)
- Adds a `EchoTo` option on the test `Writer`, for using a `StringBuilder` to dupe output and check it for assertions
- Messages on `Migrate` tests covering which end failed

Mostly this is the flags change to remediate the `MultiMaster` (heavy connection re-use, in a theory here) from #1762 and other misc. cleanup.