Bridge & Connection stats: cleanup (#1898)
We got a little heavy with stats over the years and it's cumbersome to add anything here (e.g. another `out` parameter). Instead of adding yet more stats in upcoming bits as-is, I decided to take a stab at simplifying this with `readonly struct` passes. We're not _super_ concerned with efficiency in the exception path but hey, why not.

This should simplify maintenance/additions and clarify what each property is. Note that we have some static defaults here because `default` does _not_ run property initializers where a `new()` does.

Includes a bump to C# 9 to get all the struct features in play. I tried C# 10 but the build agents didn't eat breakfast and were rather peeved I suggested it.