Change default reconnection policy to Exponential (#1896)
Change default reconnection policy to Exponential
Update the Exponential policy to take the base as parameter to allow for different growth rates

Description:
To speed up reconnections, and also avoid reconnection storms, we are changing the default reconnection policy to ExponentialRetry and reducing the deltaBackOff from the previous default of 5s to 2.5s. a
This allows more frequent retries early on, when a reconnection is more probable, and less frequent retries later on when it has already been impossible for some time.
Another plus for the ExponentialRetry is the additional randomness added to the intervals so that we avoid reconnection storms.
Here is a comparison of the two retry policies.
![image](https://user-images.githubusercontent.com/18540925/140260525-dfde3f60-2980-46e9-a0ce-1ac174f2fc80.png)
![image](https://user-images.githubusercontent.com/18540925/140260534-d137d303-e1f8-437e-a4ca-1292b9e89a4f.png)

And here is a real world example:
![image](https://user-images.githubusercontent.com/18540925/140263336-cb6b12d7-92ac-4f00-a27c-00fea2d36eaa.png)
The randomized exponential (orange) is within the the minimum delta (yellow) and the maximum exponential (gray).
The retry happens when the timeout (blue) is greater than the retry limit (orange).