Tests: handle clock skew vs. test containers (#1937)
I hit an oddity in the Docker setup locally today where the clock drifted ~10 seconds and these tests were failing because they make an inherent assumption of local clock == server clock. This aims to remove that assumption from the test suite and treat server time as accurate or account for the skew when checking TTLs.