Subscriber: simplify
This adds profiling to the test as well as fixes the `PING` going over the interactive connection as well as the `IsConnected(channel)` on `RedisSubscriber` ultimately telling us if the interactive (rather than subscription) is active - now we check both bridges.

There's still a race issue with subscriptions being restored from the `Task.Run()` in `Task.Run(() => ExecuteSubscriptionLoop());` (`GetSubscriptionQueue`) that we need to resolve, probably relying on the backlog or issuing it immediately after the handshake.